Added sort functionality

+++ src/scripts/events/domEvents.js
@@ -1,8 +1,12 @@
 import {
+  alphabetCards,
   createCard,
   deleteCard,
   filterCards,
+  getCards,
   getSingleCard,
+  newestCards,
+  oldestCards,
   showCards,
   updateCard
 } from '../cardData';
@@ -18,7 +22,7 @@ const domEvents = (uid) => {
         title: document.querySelector('#cardTitle').value,
         definition: document.querySelector('#cardDef').value,
         language: document.querySelector('#cardLang').value,
-        time: `${getTime.getHours()}:${getTime.getMinutes()}`,
+        time: getTime,
         user_id: uid
       };
       createCard(newWord).then(showCards);
@@ -32,7 +36,7 @@ const domEvents = (uid) => {
         title: document.querySelector('#cardTitle').value,
         definition: document.querySelector('#cardDef').value,
         language: document.querySelector('#cardLang').value,
-        time: `${getTime.getHours()}:${getTime.getMinutes()}`,
+        time: getTime,
         fireBaseKey,
         user_id: uid
       };
@@ -46,15 +50,27 @@ const domEvents = (uid) => {
     }
     if (e.target.id.includes('delete')) {
       console.warn('Clicked Delete Button');
-      // if (window.confirm('Are you sure you want to delete this card?')) {
-      const [, fireBaseKey] = e.target.id.split('--');
-      deleteCard(fireBaseKey, uid).then(showCards);
-      // }
+      if (window.confirm('Are you sure you want to delete this card?')) {
+        const [, fireBaseKey] = e.target.id.split('--');
+        deleteCard(fireBaseKey, uid).then(showCards);
+      }
     }
     if (e.target.id.includes('filter')) {
       const [, languageType] = e.target.id.split('--');
       console.warn(`Clicked ${languageType} Button`);
       filterCards(languageType, uid).then(showCards);
+    }
+    if (e.target.id.includes('alphabet')) {
+      console.warn('Sorting Alphabetically!');
+      getCards(uid).then(alphabetCards);
+    }
+    if (e.target.id.includes('newest')) {
+      console.warn('Sorting newest to oldest!');
+      getCards(uid).then(newestCards);
+    }
+    if (e.target.id.includes('oldest')) {
+      console.warn('Sorting oldest to newest!');
+      getCards(uid).then(oldestCards);
     }
   });
 };
